% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{check_sampleID_files}
\alias{check_sampleID_files}
\title{\\\_End_Function_\\\         #
Check existing sample ID names}
\usage{
check_sampleID_files(genepix_vars)
}
\arguments{
\item{genepix_vars}{A list of specific definitions of the experiment design.
See \code{\link{array_vars}}.}
}
\value{
A file with missing corresponding sample ID files
}
\description{
A generic function to check if the file(s) 
with the MFI values have a corresponding sample ID file. Sample ID file is
a file with the identifiers for the samples in array file.
}
\examples{
genepix_vars <- array_vars(
channel = "635",
chip_path = system.file("extdata", "array_data/machine1/", 
package="protGear"),
totsamples = 21,
blockspersample = 2,
mig_prefix = "_first",
machine = 1,
date_process = "0520"
)
check_sampleID_files(genepix_vars)
}
