% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{create_dir}
\alias{create_dir}
\title{Title Create directory function}
\usage{
create_dir(path)
}
\arguments{
\item{path}{folder location to create a directory}
}
\value{
created directory
}
\description{
creating a directory
}
\examples{
create_dir("data/sample_folder")
}
