% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{error_replicates}
\alias{error_replicates}
\title{\\\_Start_Function_For Error\\\         #}
\usage{
error_replicates(iden)
}
\arguments{
\item{iden}{An id for the file with replicates error}
}
\value{
a log file showing the replicate errors
}
\description{
A generic function to write into the log file with a replicate check error
}
\keyword{internal}
