% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_protGear_interactive.R
\name{launch_protGear_interactive}
\alias{launch_protGear_interactive}
\title{launch_protGear_interactive}
\usage{
launch_protGear_interactive()
}
\value{
launches the shiny interactive protGear app
}
\description{
This is Function is to launch the shiny application
}
\examples{
app <- system.file("shiny-examples", "protGear_interactive",
"protGear_interactive.Rmd", package = "protGear")
 if (app!=""){
 ## run this
 #launch_protGear_interactive()
 }

}
