% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{minpositive}
\alias{minpositive}
\title{Get the minimum positive value}
\usage{
minpositive(x)
}
\arguments{
\item{x}{A numeric vector or variable}
}
\value{
Returns the minimum positive value in an object
}
\description{
Get the minimum positive value
}
\examples{
minpositive(c(-1,-2,3,5,6,7,8,9,10))
}
