% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{name_of_files}
\alias{name_of_files}
\title{Object names of a list}
\usage{
name_of_files(i)
}
\arguments{
\item{i}{- a list filenames with .txt or .gpr extension}
}
\value{

}
\description{
A generic function returning a vector with the names of files in the same directory
}
