% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{plot_bg}
\alias{plot_bg}
\title{Plot background}
\usage{
plot_bg(df, x_axis = "antigen", bg_MFI = "B635_Median", log_mfi = T)
}
\arguments{
\item{df}{A     default dataset to use for plot.}

\item{bg_MFI}{A numeric \code{variable} describing which is the background MFI}

\item{log_mfi}{a logical value indicating whether the MFI values should be log transformed or not.}

\item{antigen_name}{The \code{variable} describing which features/proteins/antibodies in the data should be used to plot}
}
\value{



}
\description{
A generic function for plotting of R objects.

A generic function for plotting of R objects.
}
\details{
Plot background values
}
\examples{


}
