% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{plot_mean_sd}
\alias{plot_mean_sd}
\title{Plot mean vs SD of normalised data}
\usage{
plot_mean_sd(exprs_normalised)
}
\arguments{
\item{exprs_normalised}{A normalised object of class data frame or matrix}
}
\value{

}
\description{
A genereic function to plot \code{mean} vs \code{sd} after normalisation.
}
