% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{read_array_files}
\alias{read_array_files}
\title{Read array files}
\usage{
read_array_files(i, data_path, genepix_vars)
}
\arguments{
\item{i}{The name of the file which the data are to be read from.}

\item{data_path}{The path where the file with the data  is located}

\item{genepix_vars}{A list of specific definitions of the experiment design. See \code{\link{array_vars}}.}
}
\value{
a number of dataframes in the global environment
}
\description{
This helps to read the chip file(s).
}
\details{
Read mutiple array files
}
\examples{


}
