% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{read_array_visualize}
\alias{read_array_visualize}
\title{Read a gpr file to visualize}
\usage{
read_array_visualize(infile)
}
\arguments{
\item{infile}{a .gpr file to be used to visualize the expression intensities of the slide spots}
}
\value{
a data frame to visualize the background or foreground values
}
\description{
Read a gpr file to visualize
}
