% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{remove_batch_effect}
\alias{remove_batch_effect}
\title{Remove batch effect under development}
\usage{
remove_batch_effect()
}
\value{
a data frame of batch corrected data by limma
}
\description{
Remove batch effect under development
}
