% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{return_files}
\alias{return_files}
\title{Return files in a folder}
\usage{
return_files(filenames, data_path)
}
\arguments{
\item{filenames}{A list with the required file names to be read.}

\item{data_path}{A directory path for the file names to be read.}
}
\value{

}
\description{
This returns the list of all the files as data frames.
}
\examples{

}
