% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{rlm_normalise_matrix}
\alias{rlm_normalise_matrix}
\title{Nomrmalise using RLM}
\usage{
rlm_normalise_matrix(matrix_antigen, array_matrix, control_antigens)
}
\value{

}
\description{
A function for \code{method='rlm'} from \code{\link{matrix_normalise}}.
}
