% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{visualize_slide}
\alias{visualize_slide}
\title{Visualize the slide mimicking the original scan image.}
\usage{
visualize_slide(infile, MFI_var, interactive = F, d_f = NA)
}
\arguments{
\item{infile}{a .gpr file to be used to visualize the expression intensities of the slide spots}

\item{MFI_var}{the MFI variable to plot, can be either the background or foreground value}

\item{interactive}{a logical to specify whether an interactive graph is returned or not}
}
\value{

}
\description{
Visualize the slide mimicking the original scan image.
}
