% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{visualize_slide_2d}
\alias{visualize_slide_2d}
\title{Visualize the slide mimicking the original scan image using a 2d plot.}
\usage{
visualize_slide_2d(infile, MFI_var, d_f = NA)
}
\arguments{
\item{infile}{- a .gpr file to be used to visualize the expression 
intensities of the slide spots}

\item{MFI_var}{the MFI variable to plot, can be either the 
background or foreground value}

\item{d_f}{a data frame with array data}
}
\value{
A 2d plot of either the background or foreground values
}
\description{
Visualize the slide mimicking the original scan image using a 2d plot.
}
\examples{
## Not run:
visualize_slide_2d(
infile = system.file("extdata", "/array_data/machine1/KK2-06.txt", 
package="protGear"),
MFI_var = "B635 Median"
)
## End(Not run)
}
