% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{extract_bg}
\alias{extract_bg}
\title{extract bg}
\usage{
extract_bg(iden, data_files, genepix_vars = genepix_vars)
}
\arguments{
\item{iden}{A character indicating the name of the object to be used under data_files.}

\item{data_files}{A list of data objects with names utlised by iden.}

\item{genepix_vars}{A list of specific definitions of the experiment design. See \code{\link{array_vars}}.}
}
\value{

}
\description{
A generic function to extract the background data for micro array data.
}
\details{
\\\_End_function\\\         #

Extract the background values
}
