% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merge array data with the sample ID
\\\_Start_Function_Extract_Data\\\         #}
\usage{
merge_datasets(
  dfs,
  filenames,
  data_files,
  totsamples,
  blockspersample,
  sampleID_path,
  bg = FALSE
)
}
\arguments{
\item{dfs}{A character vector with the names of the}

\item{bg}{}
}
\description{
Merge array data with the sample ID
\\\_Start_Function_Extract_Data\\\         #
}
