% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_functions.R
\name{merge_sampleID}
\alias{merge_sampleID}
\title{Merge sample ID with the array data}
\usage{
merge_sampleID(iden, data_files, genepix_vars, method)
}
\arguments{
\item{iden}{A character indicating the name of the object to be used under data_files.}

\item{data_files}{A list of data objects with names utlised by iden.}

\item{genepix_vars}{A list of specific definitions of the experiment design. See \code{\link{array_vars}}.}

\item{method}{A description of the background correction to be used. See \code{\link{bg_correct}}.}
}
\value{

}
\description{
A generic function that merges the protein data with the sample identifiers or sample names. If the file
does not have sample identifiers the function generates it automatically.
}
