% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{output_trend_stats}
\alias{output_trend_stats}
\title{Trend test using Cox–Stuart (C–S) and Mann–Kendall (M–K) trend tests}
\usage{
output_trend_stats(name, p_val, z_val)
}
\arguments{
\item{name}{Name of the test}

\item{p_val}{p value from the test}

\item{z_val}{the Z value of the test}
}
\value{
A statistics of mean standard deviation trend
}
\description{
Trend test using Cox–Stuart (C–S) and Mann–Kendall (M–K) trend tests
}
