% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{plot_normalised}
\alias{plot_normalised}
\title{Comparison of normalised data by sample}
\usage{
plot_normalised(exprs_normalised_df, method, batch_correct)
}
\arguments{
\item{batch_correct}{}
}
\value{

}
\description{
Comparison of normalised data by sample
}
