% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{plot_normalised_antigen}
\alias{plot_normalised_antigen}
\title{Comparison of normalised data by feature}
\usage{
plot_normalised_antigen(exprs_normalised_df, method, batch_correct)
}
\arguments{
\item{batch_correct}{}
}
\value{

}
\description{
Comparison of normalised data by feature
}
