% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{rlm_normalise}
\alias{rlm_normalise}
\title{RLM normalisation}
\usage{
rlm_normalise(rlm_normalise_df)
}
\arguments{
\item{rlm_normalise_df}{rlm normalised data frame}
}
\value{
an elist of RLM normalisation to be utilised by \code{\link{rlm_normalise_matrix}}
}
\description{
A function for \code{method='rlm'} from \code{\link{matrix_normalise}}.
}
\examples{

}
